# finlogikk/front/src/app/entities/Comment.ts
import { Post } from "./Post";

export class Comment {
    idComment: any;
    content: any;
    likes: any;
    dislikes: any;
    creationDate: Date;
    modified: any;
    userId: any;
    user: any;
    post: Post;
  }