# finlogikk/front/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "yassine",
      "email": "atou26.ag@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/post/updatePost/{postId}": {
      "put": {
        "tags": [
          "post-controller"
        ],
        "operationId": "updatePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/comment/updateComment/{commentId}": {
      "put": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "updateComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
...
    "/topic/addTopic/{userId}": {
      "post": {
        "tags": [
          "topic-controller"
        ],
        "operationId": "addTopic",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Topic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    },
    "/post/addPost/{userId}/{idTopic}": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "operationId": "addPost",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idTopic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
    "/comment/addComment/{userId}/{idPost}": {
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "addComment",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idPost",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
...
    "/topic/postCount/{idTopic}": {
      "get": {
        "tags": [
          "topic-controller"
        ],
        "operationId": "countPostsByIdTopic",
        "parameters": [
          {
            "name": "idTopic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/topic/getTopic/{topicId}": {
      "get": {
        "tags": [
          "topic-controller"
        ],
        "operationId": "getTopic",
        "parameters": [
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopicDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/post/getPost/{postId}": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "operationId": "getPost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      }
    },
...
    "/post/commentCount/{idPost}": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "operationId": "countCommentsByIdPost",
        "parameters": [
          {
            "name": "idPost",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/post/byTopic/{topicId}": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "operationId": "getPostsByTopicId",
        "parameters": [
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comment/getComment/{commentId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        }
      }
    },
...
    "/comment/byPost/{postId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getCommentsByPostId",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/topic/deleteTopic/{topicId}": {
      "delete": {
        "tags": [
          "topic-controller"
        ],
        "operationId": "deleteTopic",
        "parameters": [
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/post/deletePost/{postId}": {
      "delete": {
        "tags": [
          "post-controller"
        ],
        "operationId": "deletePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/comment/deleteComment/{commentId}": {
      "delete": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "idPost": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "idComment": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        }
      },
...
      "Topic": {
        "type": "object",
        "properties": {
          "idTopic": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "TopicDTO": {
        "type": "object",
        "properties": {
          "idTopic": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "PostDTO": {
        "type": "object",
        "properties": {
          "idPost": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "CommentDTO": {
        "type": "object",
        "properties": {
          "idComment": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modified": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "clientCredentials": {
            "authorizationUrl": "http://localhost:9090/realms/finlogik/protocol/openid-connect/auth",
            "scopes": {}
          }
        }
      }
    }
  }
}